refactor: Nest report-notifications tools.
