+++ src/tools/index.ts
@@ -9,31 +9,27 @@ import "./canvases";
 import "./cost-alerts";
 import "./cost-reports";
 import "./costs";
-import "./create-report-notification";
 import "./create-virtual-tag-config";
 import "./current-user";
 import "./dashboards";
-import "./delete-report-notification";
 import "./financial-commitment-reports";
 import "./folders";
 import "./get-cost-provider-accounts";
 import "./get-financial-commitment-report";
-import "./get-report-notification";
 import "./get-team";
 import "./get-teams";
 import "./list-cost-integrations";
 import "./list-cost-providers";
 import "./list-cost-services";
-import "./list-report-notifications";
 import "./network-flow-reports";
 import "./provider-resources";
 import "./recommendation-views";
 import "./recommendations";
 import "./report-forecasts";
+import "./report-notifications";
 import "./resource-reports";
 import "./scenario-models";
 import "./tags";
-import "./update-report-notification";
 import "./user-feedback";
 import "./users";
 import "./workspaces";
