+++ test/tools/report-notifications/get-report-notification.test.ts
@@ -1,7 +1,7 @@
 import { type GetReportNotificationResponse, pathEncode } from "@vantage-sh/vantage-client";
 import { expect } from "vitest";
-import tool from "../../src/tools/get-report-notification";
-import { requestsInOrder, testTool } from "../../src/utils/testing";
+import tool from "../../../src/tools/report-notifications/get-report-notification";
+import { requestsInOrder, testTool } from "../../../src/utils/testing";
 
 const success: GetReportNotificationResponse = {
   token: "rprt_ntfctn_123",
