+++ test/tools/report-notifications/report-notification-selection-metadata.test.ts
@@ -1,12 +1,12 @@
 import { expect, test } from "vitest";
-import createCostAlert from "../../src/tools/cost-alerts/create-cost-alert";
-import getCostAlert from "../../src/tools/cost-alerts/get-cost-alert";
-import listCostAlerts from "../../src/tools/cost-alerts/list-cost-alerts";
-import createReportNotification from "../../src/tools/create-report-notification";
-import deleteReportNotification from "../../src/tools/delete-report-notification";
-import getReportNotification from "../../src/tools/get-report-notification";
-import listReportNotifications from "../../src/tools/list-report-notifications";
-import updateReportNotification from "../../src/tools/update-report-notification";
+import createCostAlert from "../../../src/tools/cost-alerts/create-cost-alert";
+import getCostAlert from "../../../src/tools/cost-alerts/get-cost-alert";
+import listCostAlerts from "../../../src/tools/cost-alerts/list-cost-alerts";
+import createReportNotification from "../../../src/tools/report-notifications/create-report-notification";
+import deleteReportNotification from "../../../src/tools/report-notifications/delete-report-notification";
+import getReportNotification from "../../../src/tools/report-notifications/get-report-notification";
+import listReportNotifications from "../../../src/tools/report-notifications/list-report-notifications";
+import updateReportNotification from "../../../src/tools/report-notifications/update-report-notification";
 
 test("report notification metadata targets natural-language report notification requests", () => {
   expect(listReportNotifications.name).toBe("list-report-notifications");
